feat: rendering each hotel's room > > Co-authored-by: Bruna <busouza01@gmail.com> > Co-authored-by: Luis <luisgsgarbelotti@gmail.com>
