+++ src/pages/Dashboard/Hotel/hotels/style.js
@@ -25,7 +25,7 @@ const Option = styled.button`
   height: 264px;
   padding: 16px 14px 22px 14px;
 
-  background-color: ${({ isSelected }) => (isSelected ? '#FFEED2' : '#F1F1F1')};
+  background-color: ${({ isHotelSelected }) => (isHotelSelected ? '#FFEED2' : '#F1F1F1')};
 
   border-radius: 10px;
   border: none;
@@ -96,4 +96,56 @@ const Option = styled.button`
   }
 `;
 
-export { Container, Option, InfoText };
+const ContainerRoom = styled.div`
+  width:100%;
+
+  margin-top: 30px;
+
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+`;
+
+const Rooms = styled.button`
+  width: 190px;
+  height: 45px;
+
+  border: 1px solid #CECECE;
+  border-radius: 10px;
+
+  background-color: ${({ isRoomSelected }) => (isRoomSelected ? '#FFEED2' : '#F1F1F1')};
+
+  padding: 0 10px;
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  cursor: pointer;
+
+  span {
+    font-family: 'Roboto';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 20px;
+    //line-height: 20px;
+    //text-align: center;
+
+    color: #454545;
+  }
+`;
+
+const Vacancies = styled.div`
+  &> * {
+    margin: 0;
+  }
+`;
+
+export {
+  Container,
+  Option,
+  InfoText,
+  ContainerRoom,
+  Rooms,
+  Vacancies
+};
